Registry: Use a compact structure
This is a breaking change.

The config file structure for registry changed a lot
as proposed in Bug T88043

The result format and behavior of /languagepairs
changed drastically. Now it allows all possible
combination of source and target languages.

The output of list/:tool/:from/:to  also changed
to accommodate the new config format. It also provides
support for default configuration of particular tool
for a language pair.

Bug: T88043
Change-Id: Icdc5c5c4e32d35e85447d22f87de7e5275d69316