cxserver should use valid language codes
To avoid all confusions, do not mix up valid MW content language codes and
language codes in the Wiki domains.

For example, bho is a valid content language while bh is not. bh is
used in the domain of bh.wikipedia.org

cxserver should never use language codes in the domain names.

Change-Id: I687e66967b0762b0decaec92261818748c43ce86