Lineardoc: Do not skip tags with empty content but with attributes
Those attributes are important and sometimes contain information for
parsoid to work with references or templates.

Added tests to verify lineardoc is not eating up empty tags

While fixing this, another bug noticed. isReference method in
Utils.js was not considering references in sup tag. This was causing
the links inside <sup> treated as usual page titles. Fixed that too
so that the tests pass.

Bug: T104539
Change-Id: I4cca925cb35a00fa27fd31380fae90e1c4c2767f