Adaptation code in cxserver and provide adaptation API - framework
This provides basic framwork for the adaptation of content including
machine translation. Real adaptation is to be done in follow up
commits.

A new api /translate/from/to is defined. That will provide machine
translation, adaptation of translation units. This output will be
readily usable in a target wiki domain. Previously with cxserver,
client side processing, adaptation was required to make the content
usable in target language wiki.

Bug: T170674
Change-Id: I92dba54620ba1be15c3fa3c89f7cf0888c9287ba