Refactor the single large registry file to smaller ones
The 1k lines registry.yaml and 10k lines registry.wikimedia.yaml are
too large and difficult to maintain without errors.

This commit splits those into separate files, and in addition uses
an algorithm for Yandex to build the language pairs from available
source and target languages.

The main config file is extended to support single string as a file
name reference that is then loaded by the new Config class.

The old registry.yaml and registry.wikimedia.yaml files are removed.
This is a breaking change and will not work with old configuration at all.

Bug: T145603
Change-Id: I00ff4d67541d118642c3d80d556eca325c9d7d6a