[BREAKING CHANGE] Return full HTML document in segmentedContent
This endpoint only returned the <body> previously, using hacky code to do so.
However, the full document is valuable for VisualEditor, especially
because of the <base> tag. Instead of stripping the HTML on the server,
send the full HTML to the client so it has a chance to look at the
header data before extracting the body.

Depends-On: Id6235710eaec77f8d9b3eef8abf6c5f838d9b2d4
Change-Id: I8367f9f27642a38104d7999c4f32b9bc7a3bc412