AdaptationTest: Actually test that the output is correct
These tests only tested that the adaptation process didn't
error out and returned a non-empty result. Now they also test
that the result is what we expect it to be.

Because the result contains data-cx blobs which are quite large
(due to embedded titleInfo for links), embed those in the test
data separately for readability, and add some code to inject them
back in by ID.

Also fix the test cases to assert the correct outcome, mostly
around link titles (the title attribute of an adapted link is
equal to the adapted href, not the unadapted one).

Change-Id: I12a723732b87c20d09967755ddb85182ee1ea9c3