Adapter#getAdapter: Do not ignore mismatch with explicit tags and rdfa
If tags are defined and an element could not match with any of it,
do not proceed with rdfa matching.

For example, previously, an inline image with <span> tag with
typeof="mw:Image/Frameless" will match MWImage translation unit
eventhough the tags does not define span. It defines only <figure>

Tests corrected accordingly.

Change-Id: Ic39cf1f1aa89137407b62396887761e5db9cbab6