Make tools API endpoint work when only source language is provided
In tools API endpoint v1/list/{tool}{/from}{/to}, 'from' and 'to'
fields are marked as optional. If none is given, or both are given,
API is working correctly. However, giving only 'from' parameter
should also work.

For a given 'from' parameter, return object of language pairs as keys,
which index arrays of supported MT providers for a language pair.

Bug: T194660
Change-Id: Ic61a9393a71b4495f2f34d3028af939ae4bdbe57