Decouple segmentation and setting link ids
If there is no segmentation to do, even then for the MW internal
links set the link id and cx-link class. Not having these attributes
will cause CX2 to consider them as non-cx links

This change caused many tests fail because the old test data we had
did not had re="mw:WikiLink" attribute for <a> tags. I could work
around that, but decided to update tests so that it is more close to
the real data we get from parsoid. But that turns out lot of file
changes.

This affects CX1 too.

Bug: T195768
Change-Id: Iae2dbca1a002a15650c7797fac298e473eebce7e