Define figure-inline as inline tag for the parser
Considering it as block tag causes segmentation errors and if
that error happens in a transclusion context, transclusion
is broken.

The consequence is that the adapters won't get the tag tree
for figure-inline like figure. That is, it wont get the img,
a tags inside figure-inline.

At the same time MWImage was not adapting these img or a tags
anyway.

Added tests from the first paragraph of en:Paris article

Can be tested in CX too. Load the article en:Paris, try translating
the first paragraph using Apertium. There should not be any
js errors, and there should be only one template for the pronunciation.

Bug: T198941
Change-Id: I016263e87aa1c339a3166c8ad8ffdd3f546b70ad