Templates: Improve the template parameter extractor
The code we used from MediaWiki:Gadget-TemplateParamWizard.js was
looking for {{{..}}} matches, which reported lot of false positives
from template code. We now looking for the left side of assignment
operator to get parameter name in source code of Template.

For example, in this source code of template:
https://ca.wikipedia.org/w/index.php?title=Plantilla:Infotaula_persona&veaction=editsource&action=edit
searching for {{3,}(.*?)[<|}] returns 875 results, while the new regex
^\|\s*([\S]+?)\s*=[\W]*{{2,} returns 219 results

Bug: T199308
Change-Id: If0b2d477782900d6bf2153ae4b329a1322e7436b