Increase the maximum request body size to 0.5mb
Default is 100 kilobytes. Sections that are larger than this size is
not unusual.

Note that the time taken for processing such a huge section depends
number or links, templates in it. It is even possible that the request
timeout. Improving this will require a non-trivial performance analysis
of cxserver that is not done in a planned way so far.

Bug: T202283
Change-Id: I159cad0930582ddc54cf5a637b9f39a4685f89a3