Fix the broken API request caching
Since MWApiRequestManager instances are not just one, the instance
cache for title pair, title info and template data are recreated
many times, effectively resulting no caching of the titles and result.

Keep the instance cache as a static for MWApiRequestManager to resolve
this issue.

Wherever Map is used, use the get, set, has, clear methods of that.

Tried to use the real ApiRequest class instances for mock data loading.

Bug: T203551
Change-Id: I92c2546ee3257bde2112a178b1b4f6ab29c1730b