Compact HTML that is sent to machine translation
Reduce the HTML payload size for MT engines by temporarily removing
all attributes except id. Attributes have no importance for an MT
engine. The attributes are then re-applied in to the MT output.

Currently Yandex is the only HTML MT engine. So the reduce/expand is
done in Yandex#translate. If we see more such MT engines in the future,
the common code can be refactored.

For example: As per T204668 a paragraph with three sentences and four
references had total 13 KB and was rejected by Yandex since it is
greater than the maximum allowed length. After this patch the payload
sent to Yandex for the same paragraph is only 2 KB.

A script to test this is also added

Bug: T204668
Change-Id: I9c3d5a30d084cfd0ee41c26c09172fa4d712fc7f