Template adaptation: Fix unhandled promise rejection error
For a template like this, the wt value is not a valid title.
The wt param value is causing "Invalid title" error from
BatchedAPIRequest.prototype.get method

 {
    "parts": [
        {
            "template": {
                "target": {
                    "wt": "formatnum: {{Infobox Finnish Municipality/population count|Vantaa}}",
                    "function": "formatnum"
                },
                "params": {},
                "i": 0
            }
        }
    ]
 }

I still don't know how to adapt this kind of templates. This commit
only catch the error and mark the template not adapted. Hence the entire
adapt api request does not fail.

Bug: T205420
Change-Id: I88cb9a10651ae9e0f7bc81173d0e8d1e3db27329