Doc#expand: Remove externally introduced attributes
Attributes in the externally introduced tags or existing tags are
removed if it does not exist in the attribute from reduce step.

This was not in the original scope of these methods, but when we discovered
cheap sanitizing feature here, make the expand process very strict.

Also add tests for that.

Change-Id: I96f69591167e3f4c302f3c6a53380dc80d76d647