Remove 'Italic title' template from article content
This template has an interesting prupose to display the article
title in italics:

"This template italicizes page titles. Article titles cannot contain
wiki-formatting, such as '', so article titles cannot be italicized
in the normal way. Instead, place this template in the article,
normally at the very top."

This has nothing to do with CX and there is nothing to customize or render
in the display. So removing it.

More over, it has an implementation sufficient enough to break the
section wrapping.

Bug: T216044
Change-Id: I809da648f6a31fabd5dea34bcf0169960ba2ce2d