Spec tests: Use a protected page instead of Oxygen and do not provide a rev
Due to RESTBase not storing old revisions of pages any more, it takes it
~10s to respond to the request for Oxygen for an old revision, given
that the page is rather long and complex. Due to this, CXServer checks
are alerting even though everything is fine. Therefore, choose a simpler
page and do not provide a revision ID in tests to alleviate the problem.

Bug: T222305
Change-Id: Ib9f6f09520bb24c353fb7ed13fb3d201d3aeb21f