Stop adding pagelanguage, hardcode it for now
CX2 is incorrectly using this to mean language of the wiki. Avoid
this bug by hardcoding it to that value instead of what is
returned by the API for the page language. Fixing the frontend
is left as a future task.

This also fixes a bug, where if you translate, say from Finnish to
Simple English, the links in the link card in the target column would
open in English Wikipedia instead of Simple English Wikipedia.

Bug: T224534
Change-Id: I417a9cd62616de1c936c38ae325621d972e8ec17