Target title fallbacks for red links
Adds new property targetFrom with possible values of link (via linked
page through Wikidata), label (via translated Label through Wikidata),
mt (machine translation) and source (no change).

This patch is impossible to review, because the tests were broken in
many ways. For example the new LRU caused results to be dropped from
the cache and not dumped on the files. This was not noticed, because
the check for offline was only checking for existence of batch api
type. If it existed (say `titleinfo` for `en`) it allowed requests
to go though. In the test files some parts were indented incorrectly
which means they were never even tested before now. This caused quite
a few changes. Recommending to review locally with -w to ignore white
space changes.

Some code is slightly shuffled around. Some code moved inside or
outside of a class depending whether cxutil.async was needed. Some
code was moved to another class to also cache siteinfo requests. This
has the downside of polluting the mocks even more. There is still
something weird going on with the mocks... such as none of them having
title map. I gave up after many hours to try to figure out that issue
for this patch.

Bug: T197787
Change-Id: I93f5ac4c083e97a1036d1f36ff118fcac7729434