Fixes for section wrapping for a complex case
Bug T216812 revealed a bug in the section wrapping code when a complex but
not rare section content is used.

A template having two fragments, and the second fragment being a figure-inline
causes the fragments in two different sections. And when that happens the first
fragment which is a span with no content but only with data-mw attributes are
dropped from the content. Effectively the template converted to HTML elements.

The Commonscat template in Spanish wikipedia was the example content.

The fix involved multiple steps.
1. The TextBlock#getTagForId method was buggy when one of the inline element is
   a sub document. The id from the document was not calculated. This caused a
   misinterpretation that the figure-inline is not a continuation of template
   started by the span tag.
2. A bug in the lineardoc/Builder caused duplicating the figure-inline in the
   generated document. While fixing this, it was found that tags with no text
   content can unnecessarily skipped from the builder output. So some cleanup
   and updating tests to match real wiki content had to be done.

Bug: T216812
Change-Id: Ib361b670f1fbd62ec73c3881931300d47e5714bc