Doc#wrapSections: Close the textblock sections immediately
So that we don't leave it open. If left open, an ignored fragment
might cause everything else to be added to the same section.

While fixing T205616 in I8a0ca75acd8dd825f5714c2e6419e64c04657ac5
we had closed the sections for self closing tags, but in this patch
we do it for all textblocks which are are not already in a section.

Tests added

Bug: T230200
Change-Id: I50482ba199e53f6b0b69fdc441cc99ef14ec0900