Integrate template parameter alignments
A machine learning approach was applied (T221211) to identify
the mappings/alignments of parameters for the most used templates
and language pairs.

The output of alignment tool is json files per language pair.
That data is loaded to an sqlite database.

This data is used as the last resort to find a corresponding param
in target template. A threshold of 0.56 is used now. The alignement
system already put a cut off of 0.45 as distance in vector space.
The score is 1-distance. So, a value of 0.55 means poor mapping.
A value of 1 means very good mapping.

Database included. All files from
https://github.com/digitalTranshumant/templatesAlignment/tree/master/templatesAligned
is loaded to database.

Test included. Mock files had to update.

Bug: T224721
Change-Id: I6a94479fcbbcb30029c698c7e0275895f01d8098