LinearDoc: Do not allow to remove empty spans in expansion step
In reduce step, if we remove all attributes-especially the typeof
attribute, the expand step will just remove inline tags with empty
content. But if we know that it is inline empty tag, but with important
attributes, we don't remove it.

An example:
<span typeof="mw:Transclusion' about="mwt#2" data-mw="..."></span>
becomes
<span id="4"></span> in reduced doc.
When we get MT output of that reduced doc, we parse it and while
parsing we see it is just a span with no content and remove it.
As a result the expanded document will miss the template.

To fix this, in reduce step, we now preserve typeof and about
attributes. The trade off is the reduced document is a bit bigger.

Test added.

Bug: T217585
Change-Id: I2cf3af371e3100e76fcbca30f2a7362f4acc1b62