Do not consider style with transclusion content as textblock
Content like:
<style typeof="mw:Extension/templatestyles mw:Transclusion"...>
 ... css goes here
</style>

used to be considered as block level content, it is true, but false
if the style contain template as in this example.

If we consider style as textblock, a sentence segmentation will happen,
isolating this tag with rest of template fragments, causing effective
template corruption.

Test case added.

This does not fix bug T217585 completely. There is an issue of templates
disappearing in the output due to a bug in reduce-expand step. Fix for
that is in I2cf3af371e3100e76fcbca30f2a7362f4acc1b62

Bug: T217585
Change-Id: Iebdeac3df817c3148d25188baa3e6d804bd9ee0d