Add mw-section-number data attribute to distinct sections in an article
In order to make a distinction where one section starts and
which paragraphs it contains, introduce new data attribute
called mw-section-number for every sections in segmented content.

Note that mw-section is different from 'sections' - the term
we used to annotate translation units.

We need to know the mw-section-number so that we can select it for
"section translation"

Added test.

Bug: T234323
Change-Id: Id0988f99690216837e91ff8ef79273d49173be1a