TextBlock#getRootItem: Consider non-whitespace text
This method tries to find a representative tag for a textblock.
Usually, that is the first tag in a textblock. But while looking
for that, if there is non-whitespace text, we cannot consider the
first tag in the content as root item.

Add tests.

For T230195, we were skipping some paragraphs from MT by wrongly
identifying them as math nodes even though there were text before
math tags to translate.

Bug: T230195
Change-Id: I799d45162d14f1a90ea9c1fa22b48c1d18e6e5c6