section suggestions: Fix the broken page metadata api
Since the metadata REST api was removed, we had to find a new API
that can give section titles.

The closest one is https://en.wikipedia.org/api/rest_v1/#/Mobile/getSections

That API is also not marked as stable, but since it is heavily used
for wikipedia mobile reading experience, I hope that we will know
if it is getting removed.

I don't want to parse the whole article DOM to get this section titles
in cxserver.

Bug: T248578
Change-Id: If4d69a751643b9e5cddd09fef3cbc9d0a0726893