Fix cxserver shutdown issue when mtClient not found
When getMTClient function fails to find a suitable MT client,
it sends a http response with the appropriate error status
code, but then translate function tries to send another http
response on top of that which leads to an
ERR_STREAM_WRITE_AFTER_END error (and cxserver shuts down).
Fixed that by returning when no mtClient is found.

Change-Id: I606b79f59599cf3f810d45035f590dc465bb03c8