Create Dockerfile to use with ContentTranslation
As part of T249901 we are trying to dockerize a system with Mediawiki,
ContentTranslation, VisualEditor and cxserver. Since docker-compose
schema version 3 does not supports 'extends' keywords, the only
way to refer other docker images in relative directories by build
contexts pointing to a Dockerfile.

For this, split the docker-compose.yml in cxserver to have Dockerfile
for the big setup and keep docker-compose.yml for standalone uses.

Change-Id: Id92476805052c8a8e27f35a099c58ce9528af4b9