Update section title mapping database based on latest CX Corpus
Based on https://dumps.wikimedia.org/other/contenttranslation/20210319/

Script used for generating the database:
https://github.com/santhoshtr/cxsectiontitlemapping

Total title pairs: 680698

Notable change: Database has index for the titles table. This improved
the table preparation time from 8 hours to 1.5 hours

Avoid rejecting suggestions that has frequency >= 100 even if that
is less than half of max-frequency suggestion. This is mostly arbitrary,
but based on some observation of actual data. For example, for en->es,
References is Referencias with frequency 600+ in the database. Bibliografía
also appears with 250+ frequency. So it is not a good idea to leave
Bibliografía here.

Noted a wrong description for the corresponding test. Corrected it.

Bug: T279045
Change-Id: Id2b0c313962594dadc12ca717371d5d79c217e35