Reference adaptaton: Support named references
Named references are references without definition but with a name
of another reference with definition. References reused in this way
were wrongly marked as not adapted by cxserver. Non-adapted references
are removed from content to publish by CX frontend. This was
causing missing references issue.

Since the reference in the context has no definition, cxserver cannot
mark the references as fully adapted. So in this patch, we are marking
them as partially adapted which are not skipped/removed by CX publishing.

Tests included.

Bug: T282369
Change-Id: I06aacbb2e345fed7a8a4f150002f8c0bfd4d6418