MT: Fix ignorable block template check
If the section is identified as ignorable for MT, we won't pass the
content to MT provider and save an API request. For example, block level
templates like infobox are does not need to be machine translated.

If the template definition is in a textblock like style, span and
the template rendering also has an non-ignorable block deep inside the
template rendering, the entire section was calculated as non-ignorable
for MT.

Fix this logic and add tests.

This issue was happening for HTML translation capable MT service like
Google MT. Sending the template HTML was returning non-parsable result.

In the test added, which is the HTML for Template:infobox writer from
en:Meg Johnson article, the block template definition is in a style tag.
Along with this style tag(a textblock), there is an img tag deep inside
the template rendering which is also a textblock but not ignorable. This
was causing the overall section non-ignorable for MT.

Bug: T309161
Change-Id: I3626f05f9caa63beebb172c155553d73e217382b