Support Audio, Video adaptation, Refactor image adaptation
An upcoming change in MediaWiki DOM Spec will use mw:File as rdfa type
attribute for images, audio and video. So far, in cxserver we are not
supporting video and audio adaptation. So the audio,video nodes will
get processed by MWImage adapted and it will break.

To avoid this, create an adapter for <img>, <audio>, <video> elements
called MWFile. This adapter will be just changing the namespace of commons
file resource to target language. Keep the MWImage adapter to adapt the
figure wrapper of images - It changes alignment classes, adapts the link
target namespace if present.

The data-cx attribute will be present in the adapted nodes indicating status.
Previously, we had the resource URLs for source and adapted target images. But
only the adaptation status is used in frontend. So just have `adapted` boolean
property.

Tests are added, updated accordingly

Change-Id: I97de9271e7f43ab9eb94728f18731a385c6f04cc