Images: Use the resource attribute of image as href for wrapped links
A regression from I97de9271e7f43ab9eb94728f18731a385c6f04cc
caused the href of <a> tags in images using src attribute of <img> tag
instead of resource attribute. If href of <a> and resource of <img>
differs, the resulting wikitext will have link= in it.

Update tests to cover this scenario.

Bug: T314836
Change-Id: I29a95b2ad0bd5deafdc404bec14ec1c7bdd8a756