MT performance: cache the translations
Check if the text content is translated in the same request.

When adapting content with links and annotations, as part of
rich text translation, many fragments(example: link text) will
be already individually translated. Later when adapting links
the link titles will require translation and it can just reuse
the previous translations.

Google and Yandex - the MT engines supporting HTML translation
would not benefit from them as we do not individually translate
annotations.

Change-Id: I0d91ca53b53285fda15f158c402b3bd54ef9b5d4