MT: guard against a malformed lineardoc while expanding html
During machine translation using untrusted services, there is a reduce-expand
step where we send very minimal html to MT service, keeping most of html tag
attributes in our side. After MT, the translated html will be expanded with
all attributes we kept. This is for security and performance purpose.

It was noted that when there are span tags with empty content as in
<span typeof="mw:Entity"> </span>, the logic to construct a lineardoc
gets wrong and adds an empty string as textblock to doc. This later causes
js errors. Add an extra check to make sure what is getting added as textblock
is indeed textblock. The validity of the example span content is questionable,
but as we work on non-trusted html content from external MT service, it is
better to be defensive.

Bug: T285217
Change-Id: I3fead44937deeae8716a85a991e916a5a1d1903c