Normalize page title before fetching page content
This patch uses the Title.getPrefixedDBKey() method to get the
normalized page title and namespace in an underscore-separated
form for multi-word titles:
e.g. "Lionel_Messi" instead of "Lionel Messi"

Tests are added too. Also fixed 2 other tests to also give
revision parameter, that is expected but was not present.

Note: We also investigated if we could use the "normalizeTitle"
method of the MwApiRequestManager class, but it doesn't fit our
needs here, as it returns space-separated titles for multi-word
titles.

Bug: T357769
Change-Id: I8cfb24d4e7f8bd1e176f1fa9a59a9c183af1bcc1