Set the proxy to dispatcher option of fetch API
Instead of using SetGlobalDispatcher, which has side effects when
fetch without proxy need to be executed, directly set dispatcher to
the fetch api.

It was not clear from the documentation if this option is available
in node fetch in addition to the undici fetch. Tested and confirmed
that the options is present and functional.

For MT Clients that does not use proxy, set a dispatcher that
will set timeouts.

A simple setup to test this is as follows:
Setup http://tinyproxy.github.io/, configure proxy option in config.yaml as http://127.0.0.1:8888. Then test the client as:

echo "<p>Hello world</p>" | node bin/translate.js Google en es

Observe the googleapi.com in the logs of tinyproxy.

Bug: T350773
Bug: T359525
Change-Id: I8a26bc8e2ae7041969ce1923ea81b1fec3fd85ab