fix: MT fails when section has template
The templatemapping configuration was not applied to read from configuration.

This is a regression from the servicerunner related rewrite.

The Adpater, TranslationUnit, MTClient and all their subclasses
expect a full application context to be passed in the constructor.
Passing the full cxconfig is not sufficient. Tests were not doing
this. Updated related tests so that they match with the actual
invocation of the above classes.

Change-Id: Ic85dc38d25ee177da7569db6e4ed6f6d56f54698