Fix error handing in /translate routes
The existing error handler was catching the adaptation errors and not the
machine translation errors. If there is a machine translation error, the
application was crashing instead of returning a 500 error.

This was the reason why we were not seeing any error logs in logstash related
to MT eventhough MT was failing occassionally.

To test this, try an invalid MinT API URL. Translate something with MinT using
swagger UI with provider as MinT. Notice that the application crashes.

After this fix, the application should return a 500 error instead of crashing.
Additionally, the error log should be printed, the /metrics listing should have
a translate_MinT_500 entry.

Change-Id: I897e3ec88521f6edfa932652c77b7374e114c0c0