Improve error handling and logging
Since express cannot handle the promise rejections, it is mandatory to
call next(error) from error handlers in route handlers. We were not
doing that and the result is application crash. While application is
crashing this error will be in stdout and in the logger output.
Crashed application will be restarted by k8s setup. but issue remains.

Refer: https://expressjs.com/en/guide/error-handling.html
Express 5 improved this, but we are on Express 4.

Call next(error) in all error handlers. In addition to this,
add a process level error handler that logs any unhandled promise
rejections and prevent application crash.

Bug: T377966
Change-Id: Ie5caa861745c729b6f17492a774e3fc773f93995