Error handling: Skip passing the error to Express
Express documentation states that if next() is called with an error
after starting the response, the default Express error handler closes
the connection and fails the request. See:
https://expressjs.com/en/guide/error-handling.html#the-default-error-handler

In our case, calling next() causes the error to be logged in a format
that we don't support, so avoid doing this if the response has already
been sent.

Bug: T382294
Change-Id: I3929af21e2ab255b27175c4a3254b53e9cab764a