SectionSuggester: Open db connection in the same execution context
In SectionSuggester, we were trying to use the connection even before
it was opened. Use promisify to wait for the connection to open before
continuing execution.

This also ensures that the errors are caught in the main execution
context and can be caught by the middleware.

Used the "cause" property of Error object
(See: https://nodejs.org/api/errors.html#errorcause) to add the
underlying cause of the Error and log it in the middleware. We can
use it in the future to identify the original error.

Bug: T383854
Bug: T377966
Change-Id: I0b78154e46b5401ee220e1cf9c8c4786807f77ba