Avoid returning response from getMTClient method
There are instances where we continue execution even if fetching an
MTClient fails but then if we try to return a response from these
routes, it fails because the getMTClient method has already sent a
response.

I found this confusing, hence refactored the method to throw an error
instead. The caller can now decide if they want to continue processing
or return an error response.

Change-Id: I140eadfdf6358e333e7708c0e551a8fab64309cd