Fix language-domain mapping for be-tarask
The mappings in "language-domain-mapping.json" are used to map the
special language codes supported in cxserver, to domain codes that
are accepted by Wikipedia Action API, so that API requests for these
codes properly work. Now, regarding "be-tarask" language code, the
Wikipedia Action API still expects "be-x-old" code, so we need to
include that mapping in the JSON object.

On the contrary, cx doesn't support "be-x-old" and "en-simple"
language codes (you can check that they are not included in the
"languages.yaml"), so it doesn't make sense to be included as
keys in the JSON object defined in language-domain-mapping.json.

Bug: T394008
Change-Id: Ie330d3c53c81b8773b201bf57d3bcd35cc7fac3f